Move config to example

+++ build-pba.sh
@@ -1,11 +1,17 @@
 #!/bin/bash
 
+if [[ ! -e config.ini ]]; then
+   echo 'No config.ini found.'
+   echo 'You must copy config-example.ini to config.ini and edit before run.'
+   exit 1
+fi
+
 if ! which cargo gdisk mkfs.fat realpath > /dev/null; then
   echo 'No cargo, gdisk, mkfs.fat or realpath'
   exit 1
 fi
 
-EFI_FILE=EFI/BOOT/BOOTX64.efi
+GREETER_EFI_FILE=EFI/BOOT/BOOTX64.efi
 OFFSET=1048576
 SECTOR_SIZE=512
 
@@ -16,7 +22,7 @@ pushd "$(dirname "$(realpath "$0")")" || exit 1
 cargo b --release || exit 1
 
 # 1mb for gpt stuff & align,
-# 1 remaining is more than enough for our image + config + remaining gpt stuff
+# 1 remaining is more than enough for our image + config.ini + remaining gpt stuff
 dd if=/dev/zero of="$IMG" bs=1M count=2
 mkdir mnt
 
@@ -48,9 +54,9 @@ echo "Asking for sudo permissions to mount the image"
 echo
 
 sudo mount -o loop,offset="$OFFSET" "$IMG" mnt || error
-sudo mkdir -p mnt/"${EFI_FILE%/*}"
-sudo cp -r target/x86_64-unknown-uefi/release/opal-uefi-greeter.efi mnt/"$EFI_FILE" || error
-sudo cp -r config-example mnt/config || error
+sudo mkdir -p mnt/"${GREETER_EFI_FILE%/*}"
+sudo cp -r target/x86_64-unknown-uefi/release/opal-uefi-greeter.efi mnt/"$GREETER_EFI_FILE" || error
+sudo cp -r config.ini mnt/config || error
 sudo umount mnt || error
 
 rmdir mnt
